+++ moirai-iter/src/parallel.rs
@@ -17,8 +17,8 @@ mod tests;
 mod traits;
 
 pub use adapters::{
-    Chain, Chunks, Cloned, Copied, Enumerate, Filter, FilterMap, FlatMap, Inspect, Intersperse,
-    Map, MapInit, MapWith, PanicFuse, Rev, Skip, Take, Update, WhileSome, Zip,
+    Chain, Chunks, Cloned, Copied, Enumerate, Filter, FilterMap, FlatMap, Flatten, Inspect,
+    Intersperse, Map, MapInit, MapWith, PanicFuse, Rev, Skip, Take, Update, WhileSome, Zip,
 };
 pub use consumers::Reduction;
 pub use sorting::ParallelSliceMut;